third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/xmlhttprequest-overridemimetype-content-type-header-expected.txt

This tests that XMLHttpRequest overrideMimeType() properly updates the Content-Type header for the response.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL xhr.getResponseHeader("Content-Type") should be text/xml;charset=GBK. Was application/xml.
PASS successfullyParsed is true

TEST COMPLETE
